backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
anthropic==0.36.0
google-generativeai==0.8.3
chromadb==0.5.15
supabase==2.9.0
apscheduler==3.10.4
resend==2.4.0
python-jose[cryptography]==3.3.0
pydantic==2.9.2
pypdf==5.0.0
pytest==8.3.3
pytest-asyncio==0.24.0
httpx==0.27.2
